refactor: 품목 검색 화면 metrics 분리
- ItemSearchScreen의 반응형 metrics 계산을 별도 파일로 이동

- 화면 렌더링 책임과 레이아웃 수치 계산 책임 분리

- 동작과 수치 기준은 유지

Refs #147
